Refactor MCPServer protocol, Models, Extensions, OpenAPI
Structural splits and function decompositions across the protocol's
default-impl extension, Dictionary parameter extraction, JSONSchema
encoding, JSONValue construction, GenericResourceContent, and
OpenAPISpec.

MCPServer.swift (929 → 77) split into:
- MCPServer+Handlers.swift — handleMessage + handleRequest split into
  domain dispatchers (dispatchInitializationOrMetaRequest,
  dispatchToolRequest, dispatchResourceRequest, dispatchPromptRequest,
  dispatchLoggingRequest), plus handleNotification/handleResponse/
  handleErrorResponse.
- MCPServer+Initialize.swift — handleInitializeRequest +
  buildServerCapabilities.
- MCPServer+Tools.swift — handleToolCall decomposed via
  buildToolCallResponse + encodeSingleContent / encodeContentArray /
  encodeResourceContent / encodeFallbackTextContent.
- MCPServer+Resources.swift — resources/* handlers.
- MCPServer+Prompts.swift — handleCompletion decomposed via a
  ParsedCompletionRequest struct + parseCompletionRequest /
  completionResponseForResource / completionResponseForPrompt /
  completion(for:in:prefix:).
- MCPServer+Logging.swift — logging/setLevel.

Dictionary+ParameterExtraction (552 → 285) split into +Numbers, +Typed,
+Values, +Strings. extractParameter (cc 15) split into
extractKnownPrimitive / extractCaseIterable / extractDecodableDynamic.
Shared invalidArgumentType / jsonTypeDescription raised to internal.

MCPParameterInfo+JSONSchema: jsonDefaultValue (cc 25, 80 lines) split
into scalarDefaultJSONValue (further decomposed) and
collectionDefaultJSONValue (ditto) + iso8601DefaultFormatter.

JSONSchema+Codable: init(from:) (cc 18, 106) decomposed into per-type
decoders. encode(to:) (cc 12, 61) moved to JSONSchema+Encoding.swift
with per-case helpers taking small payload structs. CodingKeys lifted
to internal JSONSchemaCodingKeys; AnyCodingKey extracted to its own file.

JSONValue.init(jsonObject:) (cc 25) decomposed via JSONValue+JSONObject
with matchJSONValueOrNull / matchIntegerJSONObject /
matchUnsignedIntegerJSONObject / matchFloatingPointJSONObject /
matchContainerJSONObject.

GenericResourceContent.fromResult (cc 11) split into 4 helpers.

OpenAPISpec.init (107) decomposed via collectToolMetadata /
makePathItem / errorResponseSchema / makeInputSchema.

Access widening from private → internal for cross-file refs.

Build clean. All 415 tests pass.
